# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Crypto Trading Simulator API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type GoogleAuthRequestUserInfo = {
  id: string;
  email: string;
  name: string;
  picture?: string;
};

export interface GoogleAuthRequest {
  /** Google ID token from Expo AuthSession */
  idToken: string;
  userInfo: GoogleAuthRequestUserInfo;
}

export interface UserProfile {
  id: string;
  email: string;
  name: string;
  avatar?: string | null;
  balance: number;
}

export interface AuthResponse {
  token: string;
  user: UserProfile;
}

export interface CoinPrice {
  id: string;
  symbol: string;
  name: string;
  image: string;
  currentPrice: number;
  priceChange24h: number;
  priceChangePercentage24h: number;
  marketCap: number;
  rank: number;
}

export interface TradeRequest {
  coinId: string;
  coinSymbol: string;
  coinName: string;
  quantity: number;
  price: number;
}

export interface TransactionItem {
  id: string;
  type: string;
  coinId: string;
  coinSymbol: string;
  coinName: string;
  quantity: number;
  price: number;
  total: number;
  createdAt: string;
}

export interface TradeResponse {
  transaction: TransactionItem;
  newBalance: number;
  message: string;
}

export interface HoldingItem {
  coinId: string;
  coinSymbol: string;
  coinName: string;
  quantity: number;
  avgBuyPrice: number;
  currentPrice: number;
  currentValue: number;
  profitLoss: number;
  profitLossPercent: number;
}

export interface Portfolio {
  balance: number;
  totalValue: number;
  totalInvested: number;
  totalProfitLoss: number;
  totalProfitLossPercent: number;
  holdings: HoldingItem[];
}

export interface TransactionsResponse {
  transactions: TransactionItem[];
  total: number;
}

export type GetTransactionsParams = {
  limit?: number;
  offset?: number;
};