# lib/api-zod/src/generated/types/coinPrice.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Crypto Trading Simulator API
 * OpenAPI spec version: 0.1.0
 */

export interface CoinPrice {
  id: string;
  symbol: string;
  name: string;
  image: string;
  currentPrice: number;
  priceChange24h: number;
  priceChangePercentage24h: number;
  marketCap: number;
  rank: number;
}