# lib/api-zod/src/generated/types/googleAuthRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Crypto Trading Simulator API
 * OpenAPI spec version: 0.1.0
 */
import type { GoogleAuthRequestUserInfo } from "./googleAuthRequestUserInfo";

export interface GoogleAuthRequest {
  /** Google ID token from Expo AuthSession */
  idToken: string;
  userInfo: GoogleAuthRequestUserInfo;
}