# lib/api-zod/src/generated/types/portfolio.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Crypto Trading Simulator API
 * OpenAPI spec version: 0.1.0
 */
import type { HoldingItem } from "./holdingItem";

export interface Portfolio {
  balance: number;
  totalValue: number;
  totalInvested: number;
  totalProfitLoss: number;
  totalProfitLossPercent: number;
  holdings: HoldingItem[];
}