# lib/api-zod/src/generated/types/tradeRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Crypto Trading Simulator API
 * OpenAPI spec version: 0.1.0
 */

export interface TradeRequest {
  coinId: string;
  coinSymbol: string;
  coinName: string;
  quantity: number;
  price: number;
}