# lib/api-zod/src/generated/types/transactionsResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Crypto Trading Simulator API
 * OpenAPI spec version: 0.1.0
 */
import type { TransactionItem } from "./transactionItem";

export interface TransactionsResponse {
  transactions: TransactionItem[];
  total: number;
}