# lib/api-zod/src/generated/types/userProfile.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Crypto Trading Simulator API
 * OpenAPI spec version: 0.1.0
 */

export interface UserProfile {
  id: string;
  email: string;
  name: string;
  avatar?: string | null;
  balance: number;
}